Refactor: 코드 정리 및 UI 개선, 버튼 스타일 통일 및 불필요한 요소 제거

+++ src/features/Board/components/PostDetail/PostDetailHeader.tsx
@@ -1,8 +1,10 @@
 import { deletePost } from "@/api/postApi";
-import Button, { LinkButton } from "@/components/Button";
 import StickyWrapper from "@/components/common/StickyWrapper";
+import { Button } from "@/components/ui/button";
 import { useAuthStore } from "@/stores/authStore";
 import ArrowBackIosNewIcon from "@mui/icons-material/ArrowBackIosNew";
+import { Link, useNavigate } from "react-router-dom";
+import { toast } from "react-toastify";
 import { PostData } from "../../types/type";
 
 interface PostDetailHeaderProps {
@@ -12,22 +14,37 @@ interface PostDetailHeaderProps {
 
 const PostDetailHeader = ({ post, postId }: PostDetailHeaderProps) => {
   const { user } = useAuthStore();
+  const navigete = useNavigate();
 
   return (
     <StickyWrapper>
-      <LinkButton label={<ArrowBackIosNewIcon />} to="/board" />
-      {post.uid !== user?.uid ? null : (
-        <div className="flex gap-2">
-          <LinkButton label="수정" to="edit" />
-          <Button
-            label="삭제"
-            onClick={() => {
-              deletePost(postId);
-            }}
-            color="red"
-          />
-        </div>
-      )}
+      <div className="flex justify-between">
+        <Button size="icon" asChild>
+          <Link to="/board">
+            <ArrowBackIosNewIcon />
+          </Link>
+        </Button>
+
+        {post.uid !== user?.uid ? null : (
+          <div className="flex gap-2">
+            <Button>
+              <Link to="edit">수정</Link>
+            </Button>
+
+            <Button
+              variant="destructive"
+              onClick={() => {
+                if (!confirm("정말 삭제하시겠습니까?")) return;
+                deletePost(postId);
+                toast.success("삭제 성공"); // 성공 피드백
+                navigete("/board"); // 목록 페이지로 이동
+              }}
+            >
+              삭제
+            </Button>
+          </div>
+        )}
+      </div>
     </StickyWrapper>
   );
 };
